+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - feat(template): Added multi-tenancy options to the template. (#441, #461)
 - fix(template): adjust manual chunking configuration to avoid circular deps. (#455)
 - fix(audit): key props now respect configured property exclusions.
+- fix(audit): improper default formatting for primitive collections. (#459)
 - fix: c-admin-method now preserves newlines when displaying success messages.
 - fix: c-select not defaulting optional method parameters as clearable.
 - fix: method object parameters passed sent as null on the client are received as a default instance of the object on the server. (#456)